Reworked peer handling of 'pull requests' in the case of encrypted data file restoration.

+++ src/ciyam_core_files.cpp
@@ -4661,39 +4661,39 @@ string get_account_msg_secret( const string& blockchain, const string& password,
 
 string create_peer_repository_entry_pull_info( const string& hash )
 {
-   string retval;
-
-#ifndef SSL_SUPPORT
-   throw runtime_error( "create_peer_repository_entry_pull_info requires SSL support" );
-#else
-   string local_hash, local_public_key, master_public_key;
+   string retval, local_hash, local_public_key, master_public_key;
 
    if( fetch_repository_entry_record( hash,
     local_hash, local_public_key, master_public_key, false ) && has_file( local_hash ) )
-   {
-      string file_data( c_file_type_str_blob );
-
-      file_data += c_file_repository_meta_data_line_prefix;
-      file_data += c_file_repository_meta_data_info_type_raw;
-      file_data += '\n';
-
-      file_data += c_file_repository_public_key_line_prefix;
-      file_data += base64::encode( hex_decode( local_public_key ) );
-      file_data += '\n';
-
-      file_data += c_file_repository_source_hash_line_prefix;
-      file_data += base64::encode( hex_decode( local_hash ) );
-      file_data += '\n';
-
-      file_data += c_file_repository_target_hash_line_prefix;
-      file_data += base64::encode( hex_decode( hash ) );
-
-      // NOTE: Don't allow compression to assist with interactive testing and/or debugging.
-      retval = create_raw_file( file_data, false );
-   }
-#endif
+      retval = create_peer_repository_entry_pull_info( hash, local_hash, local_public_key, master_public_key );
 
    return retval;
+}
+
+string create_peer_repository_entry_pull_info( const string& hash,
+ const string& local_hash, const string& local_public_key, const string& master_public_key )
+{
+   string retval;
+
+   string file_data( c_file_type_str_blob );
+
+   file_data += c_file_repository_meta_data_line_prefix;
+   file_data += c_file_repository_meta_data_info_type_raw;
+   file_data += '\n';
+
+   file_data += c_file_repository_public_key_line_prefix;
+   file_data += base64::encode( hex_decode( local_public_key ) );
+   file_data += '\n';
+
+   file_data += c_file_repository_source_hash_line_prefix;
+   file_data += base64::encode( hex_decode( local_hash ) );
+   file_data += '\n';
+
+   file_data += c_file_repository_target_hash_line_prefix;
+   file_data += base64::encode( hex_decode( hash ) );
+
+   // NOTE: Don't allow compression to assist with interactive testing and/or debugging.
+   retval = create_raw_file( file_data, false );
 }
 
 string create_peer_repository_entry_push_info( const string& file_hash, const string& password, string* p_pub_key )
@@ -4730,7 +4730,8 @@ string create_peer_repository_entry_push_info( const string& file_hash, const st
    return retval;
 }
 
-void decrypt_pulled_peer_file( const string& dest_hash, const string& src_hash, const string& password )
+void decrypt_pulled_peer_file( const string& dest_hash,
+ const string& src_hash, const string& password, bool was_originally_encrypted )
 {
 #ifndef SSL_SUPPORT
    throw runtime_error( "decrypt_pulled_peer_file requires SSL support" );
@@ -4773,9 +4774,14 @@ void decrypt_pulled_peer_file( const string& dest_hash, const string& src_hash, 
          stringstream ss( file_data );
          crypt_stream( ss, priv_key.construct_shared( pub_key ) );
 
-         string hash( create_raw_file( string( c_file_type_str_blob_encrypted ) + ss.str( ) ) );
+         string hash;
 
-         if( hash != dest_hash )
+         if( !was_originally_encrypted )
+            hash = create_raw_file( string( c_file_type_str_blob ) + ss.str( ) );
+         else
+            create_raw_file( string( c_file_type_str_blob_encrypted ) + ss.str( ), true, 0, 0, dest_hash.c_str( ) );
+
+         if( !was_originally_encrypted && ( hash != dest_hash ) )
          {
             delete_file( hash );
             throw runtime_error( "cannot decrypt peer file (bad password or incorrect content)" );
